Create migration for CloudinaryImage table. Implement CloudinaryImage repository and service.
Save Cloudinary image PublicId and SecureUrl to seperate table in the database which is referenced by the ImagePublicId column in the Items table. Implement CloudinaryImage repository and service. Change PicturePublicId in Item DTOs to ImageUrl. Apply changes to previous migration so the new CloudinaryImages table is created before the Items table.